refactor: flatten test-result; share dto and entity across the app
