+++ apps/nest-backend/src/controllers/report/report.controller.ts
@@ -14,8 +14,8 @@ import { IReportDetails } from '@utils';
 import { ProjectReportService } from '../../features/project-agent/project-report/project-report.service';
 import { ProjectAbstractReportService } from '../../features/project-agent/project-abstract-report/project-abstract-report.service';
 import { Log } from '../../common/logging-interceptor/logging-interceptor.service';
-import { TestDataLayerService } from '../../features/test-result/services/test-data-layer.service';
-import { FullValidationResultService } from '../../features/test-result/services/full-validation-result.service';
+import { TestDataLayerService } from '../../features/test-result/test-data-layer.service';
+import { FullValidationResultService } from '../../features/test-result/full-validation-result.service';
 
 @Controller('reports')
 export class ReportController {
