+++ apps/nest-backend/src/features/event-inspection-pipeline/event-inspection-pipeline.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
 import { EventInspectionPipelineService } from './event-inspection-pipeline.service';
 import { InspectorModule } from '../../features/inspector/inspector.module';
-import { TestResultService } from '../../features/test-result/services/test-result.service';
+import { TestResultService } from '../test-result/test-result.service';
 import { TestResultModule } from '../../features/test-result/test-result.module';
-import { ImageResultService } from '../../features/test-result/services/image-result.service';
+import { ImageResultService } from '../test-result/image-result.service';
 @Module({
   imports: [InspectorModule, TestResultModule],
   providers: [
