+++ apps/nest-backend/src/features/event-inspection-pipeline/event-inspection-pipeline.service.ts
@@ -9,9 +9,9 @@ import {
 import { EventInspectionPresetDto } from '../../shared/dto/event-inspection-preset.dto';
 import { InspectorSingleEventService } from '../../features/inspector/inspector-single-event.service';
 import { ProjectAbstractReportService } from '../../features/project-agent/project-abstract-report/project-abstract-report.service';
-import { TestResultService } from '../../features/test-result/services/test-result.service';
-import { TestResult } from '../../features/test-result/entity/test-result.entity';
-import { ImageResultService } from '../../features/test-result/services/image-result.service';
+import { TestResultService } from '../test-result/test-result.service';
+import { TestResult } from '../../shared/entity/test-result.entity';
+import { ImageResultService } from '../test-result/image-result.service';
 @Injectable()
 export class EventInspectionPipelineService {
   private readonly logger = new Logger(EventInspectionPipelineService.name);
