+++ apps/nest-backend/src/infrastructure/os/image/image.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { PathModule } from '../path/path.module';
 import { ImageService } from './image.service';
-import { ImageResultService } from '../../../features/test-result/services/image-result.service';
+import { ImageResultService } from '../../../features/test-result/image-result.service';
 import { TestResultModule } from '../../../features/test-result/test-result.module';
 
 @Module({
