+++ apps/nest-backend/src/shared/dto/configuration/create-configuration.dto.ts
