+++ apps/nest-backend/src/shared/dto/configuration/update-configuration.dto.ts
