+++ apps/nest-backend/src/shared/dto/image-result/create-image-result.dto.ts
