+++ apps/nest-backend/src/shared/dto/image-result/image-result-response.dto.ts
