+++ apps/nest-backend/src/shared/dto/image-result/update-image-result.dto.ts
