+++ apps/nest-backend/src/shared/entity/configuration.entity.ts
