+++ apps/nest-backend/src/shared/entity/image-result.entity.ts
