+++ apps/nest-backend/tsconfig.app.json
@@ -5,12 +5,10 @@
     "module": "commonjs",
     "types": ["node"],
     "emitDecoratorMetadata": true,
-    "target": "es2021"
+    "target": "es2021",
+    "composite": true,
+    "declaration": true
   },
   "exclude": ["jest.config.ts", "src/**/*.spec.ts", "src/**/*.test.ts"],
-  "include": [
-    "src/**/*.ts",
-    "../../libs/utils/src/lib/interfaces/dataLayer.interface.ts",
-    "../../libs/utils/src/lib/functions/utils.ts"
-  ]
+  "include": ["src/**/*.ts", "../../libs/utils/**/*"]
 }
